Install librealsense headers in the image
Based on a commit of same name on 'meta-intel-aero-librealsense'.

```
This change installs librealsense headers in target
rootfs so that users can compile an application that
uses librealsense on target board itself.

Signed-off-by: Pranav Tipnis <pranav.tipnis@intel.com>
```

But instead of editing librealsense recipe on that layer, do it as a
bbappend on this layer to make it easier to keep track of upstream
'meta-intel-realsense' on that layer.